public/proyectos/3-IA-Projects/claudia/requirements.txt
functions-framework>=3.0.0
google-cloud-firestore>=2.0.0
google-generativeai>=0.5.0
google-cloud-storage>=2.10.0
requests>=2.25.0
beautifulsoup4>=4.12.0
pillow>=10.0.0
lxml>=4.9.0
flask>=3.0.0
flask-cors>=4.0.0
coverage==7.11.0
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==7.0.0
pytest-mock==3.15.1
